Develop and test user signup and login logic left
