--- src/main/java/org/example/servlet/LoginServletView.java
@@ -1,21 +1,0 @@
-package org.example.servlet;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/login")
-public class LoginServletView extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException {
-        try {
-            request.setAttribute("headTitle", "Login");
-            request.getRequestDispatcher("/login.jsp").forward(request, response);
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new ServletException("Error loading login view", e);
-        }
-    }
-}
