--- src/main/java/org/example/servlet/RegisterServletView.java
@@ -1,22 +1,0 @@
-package org.example.servlet;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/register")
-public class RegisterServletView extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException {
-        try {
-            request.setAttribute("headTitle", "Register");
-            request.getRequestDispatcher("/register.jsp").forward(request, response);
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new ServletException("Error loading login view", e);
-        }
-    }
-}
-
